+++ src/main/java/com/yanny/ytech/compatibility/jei/BlockHitRecipeCategory.java
@@ -11,7 +11,6 @@ import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.category.IRecipeCategory;
-import mezz.jei.api.registration.IRecipeCatalystRegistration;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
@@ -68,9 +67,5 @@ public class BlockHitRecipeCategory implements IRecipeCategory<BlockHitRecipe> {
 
     public static List<BlockHitRecipe> getRecipes(@NotNull RecipeManager recipeManager) {
         return recipeManager.getAllRecipesFor(YTechRecipeTypes.BLOCK_HIT.get()).stream().toList();
-    }
-
-    public static void registerCatalyst(@NotNull IRecipeCatalystRegistration registration) {
-        registration.addRecipeCatalyst(new ItemStack(Items.STONE), RECIPE_TYPE);
     }
 }