+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## []
 
 - Added Bone Needle and changed leather armor and saddle recipe
+- Added Unlit Torch and changed recipe for Torch, Soul Torch, Campfire and Soul Campfire
 
 ## [0.5.0]